refactor: switch to wast
