+++ lib/WebpackOptionsDefaulter.js
@@ -73,11 +73,6 @@ class WebpackOptionsDefaulter extends OptionsDefaulter {
 			{
 				test: /\.wasm$/i,
 				type: "webassembly/experimental"
-			},
-			{
-				test: /\.wast$/i,
-				loader: "@webassemblyjs/wast-loader",
-				type: "webassembly/experimental"
 			}
 		]);
