+++ test/TestCases.template.js
@@ -141,6 +141,11 @@ const describeCases = config => {
 										{
 											test: /\.pug/,
 											loader: "pug-loader"
+										},
+										{
+											test: /\.wat$/i,
+											loader: "wast-loader",
+											type: "webassembly/experimental"
 										}
 									]
 								},
