+++ test/cases/wasm/import-wasm-wasm/index.js
@@ -1,5 +1,5 @@
 it("should allow to run a WebAssembly module with imports", function() {
-	return import("./wasm.wasm").then(function(wasm) {
+	return import("./wasm.wat").then(function(wasm) {
 		const result = wasm.addNumber(20);
 		expect(result).toEqual(42);
 	});
