+++ test/cases/wasm/import-wasm-wasm/wasm2.wat
@@ -1,0 +1,5 @@
+(module
+  (type $t0 (func (result i32)))
+  (func $getNumber (export "getNumber") (type $t0) (result i32)
+    (i32.const 22)))
+
