+++ test/cases/wasm/imports/index.js
@@ -1,5 +1,5 @@
 it("should allow to run a WebAssembly module with imports", function() {
-	return import("./wasm.wasm?1").then(function(wasm) {
+	return import("./wasm.wat?1").then(function(wasm) {
 		const result = wasm.addNumber(3);
 		expect(result).toEqual(11);
 	});
