+++ test/cases/wasm/memory/memory.wat
@@ -1,0 +1,4 @@
+(module
+  (memory $memory (export "memory") 1)
+  (data (i32.const 12) "\00\00\00\00"))
+
