+++ test/cases/wasm/memory/run.js
@@ -1,5 +1,5 @@
-import * as a1 from "./mem-access.wasm?1";
-import * as a2 from "./mem-access.wasm?2";
+import * as a1 from "./mem-access.wat?1";
+import * as a2 from "./mem-access.wat?2";
 
 a1.set(42);
 export const x1 = a1.get();
