+++ test/cases/wasm/simple/module.js
@@ -1,4 +1,4 @@
-import { add, getNumber } from "./wasm.wasm?1";
+import { add, getNumber } from "./wasm.wat?1";
 
 export function run() {
 	return add(getNumber(), 2);
