+++ test/cases/wasm/table/index.js
@@ -2,7 +2,7 @@
 const UNKNOWN_FUNCTION_TABLE = /invalid index into function table|invalid function/;
 
 it("should support tables", function() {
-	return import("./wasm-table.wasm").then(function(wasm) {
+	return import("./wasm-table.wat").then(function(wasm) {
 		expect(wasm.callByIndex(0)).toEqual(42);
 		expect(wasm.callByIndex(1)).toEqual(13);
 		expect(() => wasm.callByIndex(2)).toThrow(UNKNOWN_FUNCTION_TABLE);
@@ -10,7 +10,7 @@ it("should support tables", function() {
 });
 
 it("should support exported tables", function() {
-	return import("./wasm-table-export.wasm").then(function(wasm) {
+	return import("./wasm-table-export.wat").then(function(wasm) {
 		expect(wasm.table).toBeInstanceOf(WebAssembly.Table);
 		expect(wasm.table.length).toBe(2);
 		const e0 = wasm.table.get(0);
@@ -23,7 +23,7 @@ it("should support exported tables", function() {
 });
 
 it("should support imported tables", function() {
-	return import("./wasm-table-imported.wasm").then(function(wasm) {
+	return import("./wasm-table-imported.wat").then(function(wasm) {
 		expect(wasm.callByIndex(0)).toEqual(42);
 		expect(wasm.callByIndex(1)).toEqual(13);
 		expect(() => wasm.callByIndex(2)).toThrow(UNKNOWN_FUNCTION_TABLE);
